Add support for clearing all notifications
